+++ R/naive_Bayes.R
@@ -67,30 +67,20 @@ update.naive_Bayes <-
   function(object,
            smoothness = NULL, Laplace = NULL,
            fresh = FALSE, ...) {
-    update_dot_check(...)
+
     args <-
       list(
         smoothness = rlang::enquo(smoothness),
-        Laplace = rlang::enquo(Laplace)
+        Laplace    = rlang::enquo(Laplace)
       )
 
-    if (fresh) {
-      object$args <- args
-    } else {
-      null_args <- map_lgl(args, null_value)
-      if (any(null_args))
-        args <- args[!null_args]
-      if (length(args) > 0)
-        object$args[names(args)] <- args
-    }
-
-    new_model_spec(
-      "naive_Bayes",
-      args = object$args,
-      eng_args = object$eng_args,
-      mode = object$mode,
-      method = NULL,
-      engine = object$engine
+    update_spec(
+      object = object,
+      parameters = NULL,
+      args_enquo_list = args,
+      fresh = fresh,
+      cls = "naive_Bayes",
+      ...
     )
   }