+++ tests/testthat/test_boost_tree.R
@@ -3,20 +3,11 @@ hpc <- hpc_data[1:150, c(2:5, 8)]
 # ------------------------------------------------------------------------------
 
 test_that('updating', {
-  expr1 <- boost_tree() %>% set_engine("xgboost", verbose = 0)
-  expr2 <- boost_tree(trees = tune()) %>% set_engine("C5.0", bands = tune())
-  expr3 <- boost_tree(trees = 1, sample_size = tune())
-  expr4 <- boost_tree() %>% set_engine("C5.0", noGlobalPruning = tune())
-
-  param_tibb <- tibble::tibble(trees = 7, mtry = 1)
-  param_list <- as.list(param_tibb)
-
-  expect_snapshot(expr1 %>% update(trees = 10))
-  expect_snapshot(expr1 %>% update(param_tibb))
-  expect_snapshot(expr1 %>% update(param_list))
-  expect_snapshot(expr2 %>% update(bands = 10))
-  expect_snapshot(expr3 %>% update(trees = 1, fresh = TRUE))
-  expect_snapshot(expr4 %>% update(noGlobalPruning = TRUE))
+  expect_snapshot(
+    boost_tree(trees = 1) %>%
+      set_engine("C5.0", noGlobalPruning = TRUE) %>%
+      update(trees = tune(), noGlobalPruning = tune())
+  )
 })
 
 test_that('bad input', {