+++ NEWS.md
@@ -6,7 +6,7 @@
 
 * Exported `xgb_predict()` which wraps xgboost's `predict()` method for use with parsnip extension packages (#688).
 
-## Bug fixes
+* Fixed bug where previously set engine arguments would propagate through `update()` methods despite `fresh = TRUE` (#704).
 
 *  An inconsistency for probability type predictions for two-class GAM models was fixed (#708)